refactor: 도메인 기반 폴더 구조로 변경 및 타입 폴더 추가
