+++ src/pages/Movie/Moive.tsx
@@ -1,6 +1,6 @@
 import { Box, Container } from "@mui/material";
 import { Outlet } from "react-router-dom";
-import MovieSortButtons from "./components/BoxOffice/components/MovieSortButtons";
+import MovieSortButtons from "./components/components/MovieSortButtons";
 
 function Moive() {
   return (
