+++ src/typese/movieTypes.ts
@@ -1,9 +1,3 @@
-export interface ApiResponse<T> {
-  code: number;
-  msg: string;
-  data: T;
-}
-
 export interface IMovieDetail {
   id: number;
   title: string;
