+++ src/typese/types.ts
@@ -1,0 +1,5 @@
+export interface ApiResponse<T> {
+  code: number;
+  msg: string;
+  data: T;
+}
